# gh-109653: Improve import time of importlib.metadata / email.utils