--- a/Lib/email/_parseaddr.py
+++ b/Lib/email/_parseaddr.py
@@ -194,6 +194,7 @@ def mktime_tz(data):
         # No zone info, so localtime is better assumption than GMT
         return time.mktime(data[:8] + (-1,))
     else:
+        # Delay the import, since mktime_tz is rarely used
         import calendar
 
         t = calendar.timegm(data)