--- a/Lib/importlib/metadata/__init__.py
+++ b/Lib/importlib/metadata/__init__.py
@@ -1,7 +1,6 @@
 import os
 import re
 import abc
-import csv
 import sys
 import json
 import email
@@ -478,6 +477,8 @@ class Distribution(DeprecatedNonAbstract):
 
         @pass_none
         def make_files(lines):
+            import csv
+
             return starmap(make_file, csv.reader(lines))
 
         @pass_none