--- a/Misc/NEWS.d/next/Library/2024-01-28-00-48-12.gh-issue-109653.vF4exe.rst
+++ b/Misc/NEWS.d/next/Library/2024-01-28-00-48-12.gh-issue-109653.vF4exe.rst
@@ -1,0 +1,1 @@
+Improve import time of :mod:`importlib.metadata` and :mod:`email.utils`.