@jaraco 2024-01-28
> My criterion for delayed imports

Perhaps it would be better if there were a shared, published criteria for delayed imports, either in CPython or for Python libs in general.