@hauntsaninja 2024-01-28
@jaraco Thank you for taking the time to review. I find `timeit` hard to use to benchmark import time because of Python's caching behaviour. What timeit is doing is something like:
```
for _ in range(r):
    import sys; assert "csv" not in sys.modules
    for _ in range(n):
        import importlib.metadata
```
This explains why you see:
```
# this works
λ python -m timeit -n 1 -r 1 --setup 'import sys; assert "csv" not in sys.modules' 'import importlib.metadata'
1 loop, best of 1: 47.5 msec per loop

# but this fails
λ python -m timeit -n 1 -r 2 --setup 'import sys; assert "csv" not in sys.modules' 'import importlib.metadata'
Traceback (most recent call last):
[...]
AssertionError
```

In general, Python's caching behaviour for imports makes it very easy to benchmark incorrectly when using timeit:
```
λ cat slow.py
import time
time.sleep(1)
λ python -m timeit 'import slow'
1 loop, best of 5: 541 nsec per loop
```

> First, this change implies that best practice is not to put imports at the top of a module

I think this is a reasonable reason to reject this change. Another reasonable request to reject this change is that the benefit is relatively small and as the primary maintainer of this module you'd like a consistent style.

I think of my personal view as pragmatic: `importlib.metadata` is the slowest import in a command line tool I work on. It's unfortunate that my CLI is slower because `importlib.metadata` imports `email.utils` and `email.utils` imports `calendar` only for `mktime_tz`, a function that almost no one uses.

I agree it could be nice to have a more systemic solution. PEP 690 was a global flag to the interpreter, which complicates things, e.g. I think the SC wasn't keen on having libraries double the size of their test matrix. It's also a good point that we could put something in the dev guide about how to think about this.

> Second, the motivation for this change is lost in the commit

Thanks, I'll add a comment :-)

> Third, the importlib.metadata part of this change will need to also be contributed to importlib_metadata

I can make the same patch to `importlib_metadata` :-)