@jaraco 2024-01-28
> I can make the same patch to `importlib_metadata`. Note that of the two changes, deferral of `import calendar` in `email._parseaddr` accounts for most of the time saved — this wouldn't apply to `importlib_metadata`.

I'll take this on, since I've got the muscle memory for doing it. I just wanted to flag it for context.

> I prefer using subprocesses or general purpose benchmarking tools

Alex provided a similar suggestion in the linked issue I filed about timeit. I would like for [pytest-perf](/jaraco/pytest-perf) to be able to facilitate the testing of import latency. I'd like for pytest-perf to be one of those general purpose benchmarking tools. I've filed https://github.com/jaraco/pytest-perf/issues/12 to track that possibility.