@danielhollas 2024-01-29
>  I prefer using subprocesses or general purpose benchmarking tools

Note that there's also a command line option `-Ximporttime` specifically for disentangling the import time cost. Run e.g. like `python -Ximporttime -c "import importlib.metadata > import_time.out",
producing a text file with the import time of all the modules that end up being imported. (you might want to run it couple times to warm up disc cache, then the results are quite stable in my experience)

You can visualize it with [tuna](https://pypi.org/project/tuna/) to get a nice icicle graph of where the import time goes.