@hugovk 2024-01-29
> I've seen [PEP 690](https://peps.python.org/pep-0690/), but it appears to have been rejected (and without any reason that I could find in the PEP).

To see the SC's rejection notice, click through from "Resolution: [Discourse message](https://discuss.python.org/t/pep-690-lazy-imports-again/19661/26)" in the header; essentially they were concerned over producing a split in the community over how imports work.

> You can visualize it with [tuna](https://pypi.org/project/tuna/) to get a nice icicle graph of where the import time goes.

Yes, tuna is a nice tool. https://github.com/hugovk/pypistats/pull/289 is an example of using it to identify and improve slow imports for a CLI.

As it happens, `importlib.metadata` accounts for 27% of the final import time in that Python 3.10 example. Measuring again to zoom in, it accounts for 59% (I think because 3.12 is quicker) with 2.3% for `calendar` and 0.8% for `csv` so this PR is welcome. Thanks all!