@jaraco 2024-03-20
Backported in python/importlib_metadata@b4ce0ff90e.